# server/internal/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sirene API",
    "description": "Text-to-speech generation API with voice cloning, model management, and streaming support.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/models/{id}/pull": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Download model (SSE)",
        "description": "Starts downloading the model and streams progress via Server-Sent Events.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream with `progress`, `complete`, and `error` events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/models/{id}": {
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Delete model",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
    "/models/piper/import": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Import custom Piper model",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "onnx": {
                    "type": "string",
                    "format": "binary",
                    "description": "ONNX model file"
                  },
                  "config": {
                    "type": "string",
                    "format": "binary",
                    "description": "Piper JSON config"
                  },
                  "name": {
                    "type": "string",
                    "description": "Model name (used as slug)"
                  }
                },
                "required": [
                  "onnx",
                  "config",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Imported",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/models/{id}/export": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Export custom model as ZIP",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP archive",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}